Fixed some stuff inside linux

+++ main/libs/osmanip!ld.h
@@ -2,6 +2,6 @@
 #define LD_OSMANIP_H
 
 //This library exists so i dont need to do OS_NAME[0] everytime i want to do something thats OS dependant
-int GetTerminalSize(const char * what);
+u64 GetTerminalSize(const char * what);
 
 #endif /* LD_OSMANIP_H */
